Added hail and auto reconnect
Send HAIL on hydraStatusTopic to check if hydra is still alive. A
message on the same topic with as payload HYDRA#%timestamp%#%url% should
be send back.

Hydra will not attempt to log back in to the HomeWizard if there is a
problem with session or serial. This check is performed when a message
is being processed. Messages will be send again if reconnect was
successful.